@@ -30,6 +30,7 @@
     "@oclif/test": "^1.2.4",
     "@oclif/tslint": "^3.1.1",
     "@types/chai": "^4.1.7",
+    "@types/fs-extra": "^5.0.5",
     "@types/mocha": "^5.2.6",
     "@types/node": "^10.14.6",
     "@types/rimraf": "^2.0.2",
@@ -38,9 +39,9 @@
     "mocha": "^6.2.2",
     "nyc": "^14.1.1",
     "rimraf": "^2.6.3",
-    "ts-node": "^8.1.0",
+    "ts-node": "^9.0.0",
     "tslint": "^5.16.0",
-    "typescript": "3.5.3"
+    "typescript": "^4.0.3"
   },
   "engines": {
     "node": ">=8.0.0"
