@@ -60,8 +60,8 @@
     "rimraf": "^2.6.3",
     "source-map-support": "~0.5.16",
     "testdouble": "^3.11.0",
-    "ts-node": "^8.1.0",
+    "ts-node": "^9.0.0",
     "tslint": "^5.16.0",
-    "typescript": "^3.5.3"
+    "typescript": "^4.0.3"
   }
 }
