@@ -35,8 +35,8 @@
     "globby": "^10.0.2",
     "mocha": "^5.2.0",
     "nyc": "^14.1.1",
-    "ts-node": "^8.6.2",
-    "typescript": "^3.7.5",
+    "ts-node": "^9.0.0",
+    "typescript": "^4.0.3",
     "rimraf": "^2.6.3"
   },
   "engines": {
