@@ -61,9 +61,9 @@
     "nock": "^11.7.0",
     "nyc": "^14.1.1",
     "rimraf": "^2.6.3",
-    "ts-node": "^8.3.0",
+    "ts-node": "^9.0.0",
     "tslint": "^5.18.0",
-    "typescript": "^3.5.3",
+    "typescript": "^4.0.3",
     "uuid": "^3.3.3"
   }
 }
