@@ -72,9 +72,9 @@
     "nyc": "^14.1.1",
     "rimraf": "^3.0.0",
     "sinon": "^7.5.0",
-    "ts-node": "^8.4.1",
+    "ts-node": "^9.0.0",
     "tslint": "^5.20.1",
-    "typescript": "^3.5.3",
+    "typescript": "^4.0.3",
     "uuid": "^3.3.3"
   }
 }
