@@ -25,8 +25,8 @@
     "globby": "^10.0.1",
     "mocha": "^5.2.0",
     "nyc": "^14.1.1",
-    "ts-node": "^8.5.4",
-    "typescript": "^3.5.3",
+    "ts-node": "^9.0.0",
+    "typescript": "^4.0.3",
     "rimraf": "^2.6.3"
   },
   "engines": {
