@@ -25,15 +25,14 @@ describe('dialog:verify', () => {
         .stderr()
         .command(["dialog:verify", `${path.join(__dirname, './examples/root/*.dialog')}`, '--verbose'])
         .it('verify all', ctx => {
-            expect(ctx.stderr.match(/DLG001/).length == 7, 'Wrong number of schema errors')
-            expect(ctx.stderr.match(/DLG002/).length == 3, 'Wrong number of multiple definitions')
-            expect(ctx.stderr.match(/DLG003/).length == 1, 'Wrong number of mising definitions')
-            expect(ctx.stderr.match(/DLG004/).length == 1, 'Wrong number of type mismatches')
-            expect(ctx.stderr.match(/DLG005/).length == 6, 'Wrong number of unsued ids')
+            expect(ctx.stderr.match(/DLG001/)!.length == 7, 'Wrong number of schema errors')
+            expect(ctx.stderr.match(/DLG002/)!.length == 3, 'Wrong number of multiple definitions')
+            expect(ctx.stderr.match(/DLG003/)!.length == 1, 'Wrong number of mising definitions')
+            expect(ctx.stderr.match(/DLG004/)!.length == 1, 'Wrong number of type mismatches')
+            expect(ctx.stderr.match(/DLG005/)!.length == 6, 'Wrong number of unsued ids')
             expect(ctx.stdout).to.contain('4 files')
             expect(ctx.stderr)
                 .to.contain('Warnings: 2')
                 .to.contain('Errors: 12')
         })
 })
-
